# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: ContentAI API
  description: AI-powered content recommendation platform API
  version: "1.0.0"
  contact:
    email: admin@contentai.dev
servers:
  - url: http://localhost:4000
    description: Local development server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
        details:
          type: object
    SuccessResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
        message:
          type: string
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
          format: email
        username:
          type: string
        createdAt:
          type: string
          format: date-time
...
paths:
...
  /api/users/{id}:
    put:
      summary: Update user profile
      tags: [Users]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                email:
                  type: string
                  format: email
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/User"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/users/{id}/onboarding:
    post:
      summary: Save user onboarding preferences
      tags: [Users]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [genres, content_ids]
              properties:
                genres:
                  type: array
                  items:
                    type: string
                  example: ["Action", "Drama", "Sci-Fi"]
                content_ids:
                  type: array
                  items:
                    type: integer
                  example: [1, 5, 12]
      responses:
        "200":
          description: Onboarding saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...